# website/public/openapi.yaml
openapi: 3.1.0
info:
  title: SageReasoning API
  version: 3.0.0
  description: |
    Structured reasoning evaluation API based on Stoic philosophy.

    At its core is **sage-reason** — a universal reasoning layer that applies
    the Stoic core triad (Control Filter, Passion Diagnosis, Oikeiosis) to any
    decision input. All skills and wrappers call sage-reason internally.

    All responses use a standard envelope: `{ result, meta }`.

    Ancient reasoning, modern application. Does not consider legal, medical,
    financial, or personal obligations.
  contact:
    name: SageReasoning
    url: https://sagereasoning.com
  license:
    name: SageReasoning Proprietary Licence
    url: https://github.com/quartermileclint/stoic-brain/blob/main/LICENSE
servers:
  - url: https://www.sagereasoning.com
    description: Production
...
paths:
  /api/reason:
    post:
      tags: [sage-reason]
      summary: Universal reasoning check
      description: |
        **Outcome:** Run the Stoic core triad (or more) against any decision input.
        **Cost + Speed:** ~$0.025-$0.055 depending on depth, ~2-4s.
        **Chains To:** Any sage skill, sage-iterate, sage-guard.

        The depth parameter controls how many mechanisms are applied:
        - `quick` (3): Control Filter + Passion Diagnosis + Oikeiosis
        - `standard` (5): + Value Assessment + Appropriate Action
        - `deep` (6): + Iterative Refinement (Senecan progress tracking)
      operationId: reason
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  type: string
                  maxLength: 5000
                  description: The decision, action, or situation to reason about.
                  example: "I'm considering leaving my stable job to start a business, motivated by the potential for wealth and the fear of regret."
                context:
                  type: string
                  maxLength: 5000
                  description: Additional context for the reasoning.
                depth:
                  type: string
                  enum: [quick, standard, deep]
                  default: standard
                  description: How many Stoic mechanisms to apply.
                domain_context:
                  type: string
                  maxLength: 5000
                  description: Domain-specific context (used by skill templates).
      responses:
        '200':
          description: Reasoning evaluation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/evaluate:
    post:
      tags: [demo]
      summary: Instant demo — no auth required
      description: |
        **Outcome:** Try sage-reason instantly — no signup, no API key.
        **Cost + Speed:** Free, ~2s. Quick depth only.
        **Chains To:** /api/reason (with auth for deeper analysis).

        500 character input limit. 5 requests/minute rate limit.
      operationId: evaluate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  type: string
                  maxLength: 500
                  description: The decision to evaluate.
      responses:
        '200':
          description: Quick evaluation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeResponse'
  /api/score:
    post:
      tags: [evaluation]
      summary: 4-stage action evaluation
      description: |
        **Outcome:** Pre-action decision audit with structured reasoning + improvement path.
        **Cost + Speed:** ~$0.033, ~2s.
        **Chains To:** sage-iterate, sage-reason.
      operationId: score
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  maxLength: 2000
                context:
                  type: string
                  maxLength: 5000
                relationships:
                  type: string
                  maxLength: 5000
                emotional_state:
                  type: string
                  maxLength: 5000
                prior_feedback:
                  type: object
                  properties:
                    previous_action:
                      type: string
                    previous_proximity:
                      type: string
                    passions_identified:
                      type: array
                      items:
                        type: string
                    false_judgements:
                      type: array
                      items:
                        type: string
                    sage_reflection:
                      type: string
      responses:
        '200':
          description: Evaluation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeResponse'
...
  /api/score-iterate:
    post:
      tags: [deliberation]
      summary: Start or continue a deliberation chain
      description: |
        **Outcome:** Iterative decision refinement — submit, get feedback, revise, track improvement.
        **Cost + Speed:** ~$0.033, ~2s per iteration.
        **Chains To:** sage-score.
      operationId: scoreIterate
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                context:
                  type: string
                agent_id:
                  type: string
                chain_id:
                  type: string
                revised_action:
                  type: string
                revision_rationale:
                  type: string
      responses:
        '200':
          description: Chain step result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Supabase JWT or API key (sr_live_...)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key (sr_live_...)
  schemas:
    KatorthomaProximity:
      type: string
      enum: [reflexive, habitual, deliberate, principled, sage_like]
      description: Qualitative proximity to sage-like reasoning (no numeric scores).
    ControlFilter:
      type: object
      properties:
        within_prohairesis:
          type: array
          items:
            type: string
        outside_prohairesis:
          type: array
          items:
            type: string
    PassionDiagnosis:
      type: object
      properties:
        passions_detected:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              root_passion:
                type: string
                enum: [epithumia, hedone, phobos, lupe]
        false_judgements:
          type: array
          items:
            type: string
        correct_judgements:
          type: array
          items:
            type: string
        causal_stage_affected:
          type: string
          enum: [phantasia, synkatathesis, horme, praxis]
    Oikeiosis:
      type: object
      properties:
        relevant_circles:
          type: array
          items:
            type: object
            properties:
              stage:
                type: integer
                minimum: 1
                maximum: 5
              description:
                type: string
              obligation_met:
                type: boolean
                nullable: true
              tension:
                type: string
                nullable: true
        deliberation_notes:
          type: string
    ResponseMeta:
      type: object
      properties:
        endpoint:
          type: string
        ai_model:
          type: string
        ai_generated:
          type: boolean
        latency_ms:
          type: integer
        cost_usd:
          type: number
          nullable: true
        is_deterministic:
          type: boolean
        evaluated_at:
          type: string
          format: date-time
        composability:
          type: object
          properties:
            next_steps:
              type: array
              items:
                type: string
            recommended_action:
              type: string
        usage:
          type: object
          properties:
            monthly_calls_used:
              type: integer
            monthly_limit:
              type: integer
            monthly_remaining:
              type: integer
            usage_trigger:
              type: string
              enum: [pre_limit, at_limit]
              nullable: true
            value_summary:
              type: string
    ReasonResponse:
      type: object
      properties:
        result:
          type: object
          properties:
            control_filter:
              $ref: '#/components/schemas/ControlFilter'
            passion_diagnosis:
              $ref: '#/components/schemas/PassionDiagnosis'
            oikeiosis:
              $ref: '#/components/schemas/Oikeiosis'
            value_assessment:
              type: object
              description: Present at standard and deep depth.
            kathekon_assessment:
              type: object
              description: Present at standard and deep depth.
            iterative_refinement:
              type: object
              description: Present at deep depth only.
            katorthoma_proximity:
              $ref: '#/components/schemas/KatorthomaProximity'
            philosophical_reflection:
              type: string
            improvement_path:
              type: string
            disclaimer:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    EnvelopeResponse:
      type: object
      description: Standard response envelope used by all endpoints.
      properties:
        result:
          type: object
          description: Evaluation-specific result data.
        meta:
          $ref: '#/components/schemas/ResponseMeta'
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string